Move `fixture`s into `conftest.py`
